move `remove_none_values`
